% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R, R/components.R
\name{print.adlp_components}
\alias{print.adlp_components}
\alias{adlp_components}
\title{Accident and Development period Adjusted Linear Pools Component Models}
\usage{
\method{print}{adlp_components}(object, ...)

adlp_components(...)
}
\arguments{
\item{...}{Individual \code{adlp_components}}
}
\description{
Accident and Development period Adjusted Linear Pools Component Models
}
\details{
Class to structure a list of \link[adlp_component]{adlp_components}.
}
