% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{calc_adlp_component}
\alias{calc_adlp_component}
\alias{calc_adlp_component_lst}
\title{Accident and Development period Adjusted Linear Pools Component Models}
\usage{
calc_adlp_component(
  component,
  newdata,
  y = NULL,
  model = c("train", "full"),
  calc = c("pdf", "cdf", "mu", "sim")
)

calc_adlp_component_lst(
  components_lst,
  newdata,
  model = c("train", "full"),
  calc = c("pdf", "cdf", "mu", "sim"),
  ...
)
}
\arguments{
\item{component}{Object of class `adlp_component`}

\item{newdata}{Claims Triangle and other information. `data.frame` format of
claims and related information for each cell. Dataframe will have columns
 `origin` and `dev` as columns 1 and 2 respectively.}

\item{y}{Optional vector of `y` to be used in pdf or cdf calculations. Will
default to the response fetched by `model.frame`.}

\item{model}{Whether the training component model or the full component model
should be used}

\item{calc}{Type of calculation to perform}
}
\description{
Accident and Development period Adjusted Linear Pools Component Models
}
\details{
Calls the specified function for an object of class `adlp_component`.

`calc_adlp_component_lst` is a wrapper for `calc_adlp_component` for each
component in the list `components_lst`. This wrapper also allows signals the
component that casuses and error if occuring downstream.
}
