% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_model.R
\name{custom_model}
\alias{custom_model}
\alias{update.custom_model}
\title{Custom Model Wrapper}
\usage{
custom_model(formula, data, ...)

\method{update}{custom_model}(x, data)
}
\arguments{
\item{formula}{Formula needed that defines all variables required for the model}

\item{data}{Initial training data for model}
}
\description{
Function to define basic functionality needed for a custom model that does
not fit the general framework of models that align with \link[adlp_component]{adlp_component}
}
\details{
Custom model should support the S3 method \code{formula} and \code{update}.
}
