% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_adlp_component}
\alias{test_adlp_component}
\title{Test ADLP Component}
\format{
A \code{adlp_component} format, see \link[ADLP]{adlp_component}.
}
\usage{
test_adlp_component
}
\description{
A \code{adlp_component} object created for examples.
}
\examples{
test_adlp_component
}
\keyword{datasets}
