% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_val_split.R
\name{train_val_split}
\alias{train_val_split}
\title{Train-Validation Split of Claims Triangle}
\usage{
train_val_split(df, ...)
}
\arguments{
\item{df}{Claims Triangle and other information. \code{data.frame} format of
claims and related information for each cell. Dataframe will have columns
\code{origin} and \code{dev} as columns 1 and 2 respectively.}

\item{...}{Other parameters used to calculate train/test splitting.}
}
\value{
List containing \verb{$train}, \verb{$valid}, \verb{$test}, which should partition
the input \code{df}.
}
\description{
General framework for any user-defined training/validation/testing split of
claims triangle data. The ADLP package contains three default splitting algorithms.
}
\seealso{
\link[ADLP]{train_val_split_by_AP},
\link[ADLP]{train_val_split_method1},
\link[ADLP]{train_val_split_method2}
}
