% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_val_split.R
\name{train_val_split_by_AP}
\alias{train_val_split_by_AP}
\title{Train-Validation Split by Accident Period}
\usage{
train_val_split_by_AP(df, accident_periods, max_dev_periods, test = FALSE)
}
\arguments{
\item{df}{Claims Triangle and other information. `data.frame` format of
claims and related information for each cell. Dataframe will have columns
 `origin` and `dev` as columns 1 and 2 respectively.}

\item{accident_periods}{Vector of accident periods. Will be equivalent to
`1:Triangle_Size`}

\item{max_dev_periods}{Vector of development periods}

\item{test}{Returns the test set if `TRUE`}
}
\value{
List containing `$train`, `$valid`, `$test`, which should partition
 the input `df`.
}
\description{
Function for training/validation splitting.
}
\details{
Assigns training set defined by a maximum development period for each
accident period: \eqn{(x_{ij} <= MaxDP(i))}.

Validation set is therefore cells outside of this period but within the
upper triangle. The test set is all observations in the lower triangle.
}
\seealso{
\link[train_val_split]{train_val_split}
}
