% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_methods.R, R/components.R
\name{print.adlp_components}
\alias{print.adlp_components}
\alias{adlp_components}
\title{Accident and Development period Adjusted Linear Pools Component Models}
\usage{
\method{print}{adlp_components}(x, ...)

adlp_components(...)
}
\arguments{
\item{x}{Object of class \code{adlp_components}}

\item{...}{Individual \code{adlp_components}}
}
\value{
An object of class \code{adlp_components}
}
\description{
Accident and Development period Adjusted Linear Pools Component Models
}
\details{
Class to structure a list of \link[ADLP]{adlp_components}.
}
\examples{
data(test_adlp_component)
test_component1 <- test_adlp_component
test_component2 <- test_adlp_component
test_components <- adlp_components(
    component1 = test_component1,
    component2 = test_component2
)

}
